Seperate class for iptables
